Buggy (but working) vectorized unit processing.
- Modulation messages are passed in order (by reference), but midi
  messages are not fully implemented
- Note on and note off messages are sent, but not in correct sample
  order. They are sent at the beginning processing call. So they need to
  be sent with sample information in a way that unit.tick can discover
  incoming midi messages on the correct sample.
- Oscilloscope is broken. It needs to be able to store a buffer that is
  larger than the estimated period. It needs to be able to keep track of which
  part of the buffer is relevant to the current period and which part is
  for the next period (and then update as necessary).
- VosimChoir and RandomOscillator are slow. Maybe vectorize
  RandomOscillator to output a matrix of samples, where each row
  represents a separate RandomOscillator. Also, some random data needs
  to occasionally precomputed (perhaps during instrument downtime).

+++ OscilloscopeTransforms.cpp
@@ -65,7 +65,7 @@ namespace syn
     {
       oscconfig.outputbuf[i] = inputbuf[i];
       oscconfig.xaxisticks[i] = i/(double)N*pPlug->GetSampleRate();
-      snprintf(lblbuf, 64, "%f", i/pPlug->GetSampleRate());
+      snprintf(lblbuf, 64, "%f", i/(double)pPlug->GetSampleRate());
       oscconfig.xaxislbls[i] = string(lblbuf);
       if (oscconfig.argmax == -1 || oscconfig.outputbuf[i] > oscconfig.outputbuf[oscconfig.argmax])
         oscconfig.argmax = i;
